requirements.txt
fastapi==0.109.0
uvicorn==0.24.0
sqlalchemy==2.0.25
jinja2==3.1.3
openai==1.12.0
tenacity==8.2.3
pillow==10.2.0
python-multipart==0.0.9
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
httpx==0.28.1
requests==2.31.0
